Refactoring Fetch Applications + Fetch Favorites
This will make it easier to reuse on the Profile Page
