+++ frontend/src/Pages/Applications.js
@@ -20,6 +20,7 @@ import DeleteButton from "../components/CommonComp/Button/DeleteButton";
 import { useSearchParams } from "react-router-dom";
 import usePagination from "../Utils/usePagination/usePagination";
 import { UserContext } from "../Utils/UserContext/UserContext";
+import { FetchApplicationsHook } from "../Utils/DataFetching/FetchApplicationsHook";
 
 function Applications() {
 	const location = useLocation();
@@ -35,117 +36,14 @@ function Applications() {
 		useState(false);
 	const [applicationToView, setApplicationToView] = useState(null);
 
-	const { route, user } = useAuthenticator((context) => [
-		context.route,
-		context.user,
-	]);
-	const { getAccessToken } = useContext(UserContext);
-
-	const {
-		isError,
-		isLoading,
-		isSuccess,
-		error,
-		data: applicationData,
-		refetch,
-	} = useQuery(
-		["userApplications"],
-		async () => {
-			const accessToken = await getAccessToken();
-			return API.get("HMEBackend", `/api/user/a`, {
-				headers: {
-					Authorization: "Bearer " + accessToken || null,
-				},
-				queryStringParameters: {
-					userId: user?.username || null,
-				},
-			});
-		},
-		{
-			response: true,
-			refetchOnMount: true,
-			refetchOnWindowFocus: false,
-			enabled: true,
-			select: (data) => {
-				return data;
-			},
-			onSuccess: (data) => {
-				console.log(data);
-			},
-			onError: (err) => {
-				console.log(err);
-			},
-		}
-	);
-
-	const applicationPropIds = applicationData?.map((el) => {
-		return el.propertyId.toString();
-	});
-
-	const {
-		isError: detailsIsError,
-		isLoading: detailsIsLoading,
-		error: detailsError,
-		data: detailsData,
-		refetch: detailsRefetch,
-	} = useQuery(
-		["propertyQuickViews"],
-		async () => {
-			const accessToken = await getAccessToken();
-
-			return API.get("HMEBackend", `/api/properties/batch`, {
-				headers: {
-					Authorization: "Bearer " + accessToken || null,
-				},
-				queryStringParameters: {
-					ids: applicationPropIds,
-				},
-			});
-		},
-		{
-			response: true,
-			enabled: false,
-			refetchOnWindowFocus: false,
-			refetchOnMount: true,
-			select: (rawData) => {
-				let out = [
-					...rawData.map((el) => {
-						let appDet = applicationData.find((app) => {
-							return app.propertyId == el.propertyId;
-						});
-						return {
-							...el,
-							message: appDet.message,
-							price: el.price,
-							address: el.streetAddress,
-							thumbnail: el.images[0],
-							status: "pending",
-						};
-					}),
-				];
-
-				return out;
-			},
-			onSuccess: (data) => {
-				// console.log(data);
-			},
-			onError: (err) => {
-				// console.log(err);
-			},
-		}
-	);
+	const { detailsIsLoading, detailsData, isLoading, refetch } =
+		FetchApplicationsHook();
 
 	useEffect(() => {
 		setInitialBreadcrumbLocation(
 			location.state?.previousUrl ? location.state.previousUrl : null
 		);
 	}, []);
-
-	useEffect(() => {
-		if (isSuccess) {
-			detailsRefetch();
-		}
-	}, [applicationData, isLoading, isSuccess]);
 
 	const { pageNum, handlePageChange } = usePagination(() => {},
 	setSearchParameters);
