+++ frontend/src/Pages/Favorites.js
@@ -24,6 +24,7 @@ import { useEffect, useContext } from "react";
 import { UserContext } from "../Utils/UserContext/UserContext";
 import SkeletonCard from "../components/CommonComp/Cards/SkeletonCard/SkeletonCard";
 import PropertyCard from "../components/CommonComp/Cards/PropertyCard/PropertyCard";
+import { FetchFavoritesHook } from "../Utils/DataFetching/FetchFavoritesHook";
 
 function Favorites() {
 	const theme = useTheme();
@@ -41,14 +42,9 @@ function Favorites() {
 	const [initialBreadcrumbLocation, setInitialBreadcrumbLocation] =
 		useState(null);
 
-	const methods = useForm({
-		defaultValues: { ...DEFAULT_FAVORITE_FILTER_VALUES },
-	});
-	const { route, user } = useAuthenticator((context) => [
-		context.route,
-		context.user,
-	]);
-	const { getAccessToken } = useContext(UserContext);
+	// const methods = useForm({
+	// 	defaultValues: { ...DEFAULT_FAVORITE_FILTER_VALUES },
+	// });
 	// const filterSubmit = (formdata) => {
 	// 	setSearchParameters((params) => {
 	// 		Object.keys(formdata).forEach((key) => {
@@ -60,98 +56,7 @@ function Favorites() {
 	// 	refetch();
 	// };
 
-	const {
-		isSuccess,
-		isError,
-		isLoading,
-		error,
-		data: favoriteData,
-		refetch,
-	} = useQuery(
-		["userFavourites"],
-		async () => {
-			const accessToken = await getAccessToken();
-			return API.get("HMEBackend", `/api/user/f`, {
-				headers: {
-					Authorization: "Bearer " + accessToken || null,
-				},
-			});
-		},
-		{
-			refetchOnMount: true,
-			response: true,
-			queryStringParameters: {
-				userId: user?.username || null,
-				...methods.getValues(),
-			},
-			selector: (data) => {
-				const out = [
-					...data.data?.map((el) => {
-						return el.favourite;
-					}),
-				];
-				return out;
-			},
-			onSuccess: (data) => {
-				// console.log(data);
-			},
-		}
-	);
-
-	const favoriteIds = favoriteData
-		?.map((el) => {
-			return el.favourite.toString();
-		})
-		.filter((el) => {
-			return el != "1";
-		});
-
-	// console.log(favoriteIds);
-
-	const {
-		isError: detailsIsError,
-		isLoading: detailsIsLoading,
-		error: detailsError,
-		data: detailsData,
-		refetch: detailsRefetch,
-	} = useQuery(
-		["favoriteQuickViews"],
-		() => {
-			return API.get("HMEBackend", `/api/properties/batch`, {
-				headers: {
-					Authorization:
-						"Bearer " +
-							user?.getSignInUserSession().getAccessToken().getJwtToken() ||
-						null,
-				},
-				queryStringParameters: {
-					ids: favoriteIds,
-				},
-			});
-		},
-		{
-			response: true,
-			enabled: false,
-			refetchOnWindowFocus: false,
-			select: (data) => {
-				return data;
-			},
-			onSuccess: (data) => {
-				console.log(data);
-			},
-			onError: (err) => {
-				// console.log(err);
-			},
-		}
-	);
-
-	useEffect(() => {
-		if (isSuccess) {
-			detailsRefetch();
-		}
-	}, [favoriteData, isLoading, isSuccess]);
-
-	// methods.customSubmitBehavior = filterSubmit;
+	const { detailsData, isError, isLoading, refetch } = FetchFavoritesHook();
 
 	useEffect(() => {
 		setInitialBreadcrumbLocation(
